# swagger-static/swagger-spec.yaml
openapi: 3.0.0
paths:
...
  /api/v1/tags:
    post:
      operationId: TagsController_createOne
      summary: Создание нового тега
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagDto'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '401':
          description: Unauthorized
      tags: &ref_2
        - tags
      security: &ref_3
        - bearer: []
    get:
      operationId: TagsController_findAll
      summary: Получение списка тегов
      parameters: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '401':
          description: Unauthorized
      tags: *ref_2
      security: *ref_3
  /api/v1/tags/{id}:
    get:
      operationId: TagsController_findOneById
      summary: Получение тега по ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_2
      security: *ref_3
    put:
      operationId: TagsController_updateOneById
      summary: Обновление тега по ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTagDto'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_2
      security: *ref_3
    delete:
      operationId: TagsController_deleteOneById
      summary: Удаление тега по ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_2
      security: *ref_3
  /api/v1/tags/slug/{slug}:
    get:
      operationId: TagsController_findOneBySlug
      summary: Получение тега по slug
      parameters:
        - name: slug
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_2
      security: *ref_3
...
  /api/v1/users:
    get:
      operationId: UsersController_findAll
      summary: Получение списка пользователей
      parameters: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
      tags: &ref_6
        - users
      security: &ref_7
        - bearer: []
  /api/v1/users/{id}:
    get:
      operationId: UsersController_findOneById
      summary: Получение пользователя по ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_6
      security: *ref_7
    put:
      operationId: UsersController_updateOneById
      summary: Обновление пользователя по ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_6
      security: *ref_7
    delete:
      operationId: UsersController_deleteOneById
      summary: Удаление пользователя по ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_6
      security: *ref_7
  /api/v1/users/{username}:
    get:
      operationId: UsersController_findOneByUsername
      summary: Получение пользователя по username
      parameters:
        - name: username
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_6
      security: *ref_7
  /api/v1/teams:
    post:
      operationId: TeamsController_createOne
      summary: Создание новой команды
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamDto'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '401':
          description: Unauthorized
      tags: &ref_8
        - teams
      security: &ref_9
        - bearer: []
    get:
      operationId: TeamsController_findAll
      summary: Получение списка команд
      parameters: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '401':
          description: Unauthorized
      tags: *ref_8
      security: *ref_9
  /api/v1/teams/{id}:
    get:
      operationId: TeamsController_findOneById
      summary: Получение команды по ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_8
      security: *ref_9
    put:
      operationId: TeamsController_updateOneById
      summary: Обновление команды по ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTeamDto'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_8
      security: *ref_9
    delete:
      operationId: TeamsController_deleteOneById
      summary: Удаление команды по ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '401':
          description: Unauthorized
        '404':
          description: Not Found
      tags: *ref_8
      security: *ref_9
...
info:
  title: Студенческий код
  description: Это открытое API для веб-приложения "Студенческий код".
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    ProjectLinksDto:
      type: object
      properties:
        main:
          type: string
          description: Основная ссылка
        github:
          type: string
          description: Ссылка на GitHub
        demo:
          type: string
          description: Ссылка на демо
      required:
        - main
        - github
        - demo
...
    ObjectId:
      type: object
      properties: {}
    TagNameDto:
      type: object
      properties:
        en:
          type: string
          description: Название на английском
        ru:
          type: string
          description: Название на русском
      required:
        - en
        - ru
    Tag:
      type: object
      properties:
        _id:
          description: Идентификатор
          allOf:
            - $ref: '#/components/schemas/ObjectId'
        name:
          description: Название
          allOf:
            - $ref: '#/components/schemas/TagNameDto'
        icon:
          type: string
          description: Ссылка на иконку
        slug:
          type: string
          description: Ключевое слово
      required:
        - _id
        - name
        - icon
        - slug
    UserFullNameDto:
      type: object
      properties:
        surname:
          type: string
          description: Фамилия
        name:
          type: string
          description: Имя
        patronymic:
          type: string
          description: Отчество
      required:
        - surname
        - name
        - patronymic
    UserLinksDto:
      type: object
      properties:
        github:
          type: string
          description: Гитхаб
        vkontakte:
          type: string
          description: Вконтакте
        telegram:
          type: string
          description: Телеграм
      required:
        - github
        - vkontakte
        - telegram
    AwardNameDto:
      type: object
      properties:
        en:
          type: string
          description: Название на английском
        ru:
          type: string
          description: Название на русском
      required:
        - en
        - ru
    Award:
      type: object
      properties:
        _id:
          description: Идентификатор
          allOf:
            - $ref: '#/components/schemas/ObjectId'
        name:
          description: Название
          allOf:
            - $ref: '#/components/schemas/AwardNameDto'
        icon:
          type: string
          description: Ссылка на иконку
      required:
        - _id
        - name
        - icon
    User:
      type: object
      properties:
        _id:
          description: Идентификатор
          allOf:
            - $ref: '#/components/schemas/ObjectId'
        username:
          type: string
          description: Имя пользователя
        email:
          type: string
          description: Электронная почта
        password:
          type: string
          description: Пароль
        role:
          type: string
          description: Роль
          enum:
            - user
            - admin
        refresh_token:
          type: string
          description: Токен обновления
        full_name:
          description: ФИО
          allOf:
            - $ref: '#/components/schemas/UserFullNameDto'
        avatar:
          type: string
          description: Ссылка на аватар
        about:
          type: string
          description: О себе
        links:
          description: Ссылки
          allOf:
            - $ref: '#/components/schemas/UserLinksDto'
        awards:
          description: Награды
          type: array
          items:
            $ref: '#/components/schemas/Award'
        projects:
          description: Проекты
          type: array
          items:
            type: array
      required:
        - _id
        - username
        - email
        - password
        - role
        - refresh_token
        - full_name
        - avatar
        - about
        - links
        - awards
        - projects
    Project:
      type: object
      properties:
        _id:
          description: Идентификатор
          allOf:
            - $ref: '#/components/schemas/ObjectId'
        title:
          type: string
          description: Название
        tagline:
          type: string
          description: Слоган
        status:
          type: string
          description: Статус
          enum:
            - draft
            - published
            - archived
        description:
          type: string
          description: Описание
        flames:
          type: number
          description: Количество огоньков
        links:
          description: Ссылки
          allOf:
            - $ref: '#/components/schemas/ProjectLinksDto'
        logo:
          type: string
          description: Ссылка на логотип
        screenshots:
          description: Массив скриншотов
          type: array
          items:
            type: string
        price:
          type: string
          description: Цена
          enum:
            - free
            - free_options
            - payment_required
        rating:
          type: number
          description: Рейтинг
        tags:
          description: Теги
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        creator:
          description: Создатель
          allOf:
            - $ref: '#/components/schemas/User'
      required:
        - _id
        - title
        - tagline
        - status
        - description
        - flames
        - links
        - logo
        - screenshots
        - price
        - rating
        - tags
        - creator
...
    CreateTagDto:
      type: object
      properties:
        name:
          description: Название
          allOf:
            - $ref: '#/components/schemas/TagNameDto'
        icon:
          type: string
          description: Ссылка на иконку
        slug:
          type: string
          description: Ключевое слово
      required:
        - name
        - icon
        - slug
    UpdateTagDto:
      type: object
      properties: {}
...
    UpdateUserDto:
      type: object
      properties: {}
    CreateTeamDto:
      type: object
      properties:
        name:
          type: string
          description: Название
        about:
          type: string
          description: О команде
        logo:
          type: string
          description: Ссылка на логотип
        users:
          description: Участники
          type: array
          items:
            type: string
        projects:
          description: Проекты
          type: array
          items:
            type: string
      required:
        - name
        - about
        - logo
        - users
        - projects
    Team:
      type: object
      properties:
        _id:
          description: Идентификатор
          allOf:
            - $ref: '#/components/schemas/ObjectId'
        name:
          type: string
          description: Название
        about:
          type: string
          description: О команде
        logo:
          type: string
          description: Ссылка на логотип
        users:
          description: Участники
          type: array
          items:
            $ref: '#/components/schemas/User'
        projects:
          description: Проекты
          type: array
          items:
            $ref: '#/components/schemas/Project'
      required:
        - _id
        - name
        - about
        - logo
        - users
        - projects
    UpdateTeamDto:
      type: object
      properties: {}
...